requirements.txt
opencv_python==4.0.0.21
numpy==1.16.0
pytesseract==0.2.6
tensorflow==1.12.0
Keras==2.2.4
Pillow==6.0.0
PyQt4==4.11.4
